🗺️ Atlas: Extract cli dispatchers to resolve mod.rs bloat
