+++ tests/rehearsal_tests.rs
@@ -263,7 +263,9 @@ async fn test_skip_evaluator_short_circuit() {
         Command::Bench { cmd } => match *cmd {
             maxwells_daemon::cli::args::BenchCmd::Rehearsal(mut s) => {
                 s.rehearse = true;
-                maxwells_daemon::cli::bench_swebench(*s).await.unwrap();
+                maxwells_daemon::cli::bench::bench_swebench(*s)
+                    .await
+                    .unwrap();
             }
             _ => panic!("Expected bench rehearsal subcommand"),
         },
@@ -313,7 +315,9 @@ async fn test_diff_surfaces_drift() {
     if let Command::Bench { cmd } = cli_baseline.command {
         if let maxwells_daemon::cli::args::BenchCmd::Rehearsal(mut s) = *cmd {
             s.rehearse = true;
-            maxwells_daemon::cli::bench_swebench(*s).await.unwrap();
+            maxwells_daemon::cli::bench::bench_swebench(*s)
+                .await
+                .unwrap();
         }
     }
 
@@ -336,7 +340,9 @@ async fn test_diff_surfaces_drift() {
     if let Command::Bench { cmd } = cli_candidate_match.command {
         if let maxwells_daemon::cli::args::BenchCmd::Rehearsal(mut s) = *cmd {
             s.rehearse = true;
-            maxwells_daemon::cli::bench_swebench(*s).await.unwrap();
+            maxwells_daemon::cli::bench::bench_swebench(*s)
+                .await
+                .unwrap();
         }
     }
 
@@ -370,7 +376,9 @@ async fn test_diff_surfaces_drift() {
     if let Command::Bench { cmd } = cli_candidate_drift.command {
         if let maxwells_daemon::cli::args::BenchCmd::Rehearsal(mut s) = *cmd {
             s.rehearse = true;
-            maxwells_daemon::cli::bench_swebench(*s).await.unwrap();
+            maxwells_daemon::cli::bench::bench_swebench(*s)
+                .await
+                .unwrap();
         }
     }
 
@@ -563,7 +571,7 @@ async fn test_forecast_rejects_rehearsal() {
     let cli = Cli::try_parse_from(args).unwrap();
     if let Command::Bench { cmd } = cli.command {
         if let maxwells_daemon::cli::args::BenchCmd::Swebench(s) = *cmd {
-            let res = maxwells_daemon::cli::bench_swebench(*s).await;
+            let res = maxwells_daemon::cli::bench::bench_swebench(*s).await;
             assert!(res.is_err());
             let err_msg = res.unwrap_err().to_string();
             assert!(err_msg.contains("rehearsal mode cannot be used with forecast-first"));
@@ -808,7 +816,7 @@ async fn test_rehearsal_rejects_diff_combined_with_dry_run() {
     let cli = Cli::try_parse_from(args).unwrap();
     if let Command::Bench { cmd } = cli.command {
         if let maxwells_daemon::cli::args::BenchCmd::Swebench(s) = *cmd {
-            let res = maxwells_daemon::cli::bench_swebench(*s).await;
+            let res = maxwells_daemon::cli::bench::bench_swebench(*s).await;
             assert!(res.is_err());
             let err_msg = res.unwrap_err().to_string();
             assert!(err_msg.contains("--diff cannot be used with --dry-run"));
@@ -919,7 +927,7 @@ async fn test_rehearsal_rejects_empty_terminal_path_segment() {
     let cli = Cli::try_parse_from(args).unwrap();
     if let Command::Bench { cmd } = cli.command {
         if let maxwells_daemon::cli::args::BenchCmd::Swebench(s) = *cmd {
-            let res = maxwells_daemon::cli::bench_swebench(*s).await;
+            let res = maxwells_daemon::cli::bench::bench_swebench(*s).await;
             assert!(res.is_err());
             let err_msg = res.unwrap_err().to_string();
             assert!(err_msg.contains("rehearsal output path must contain a terminal path segment"));
